chore: update lakebase path
