+++ src/components/home/features/index.tsx
@@ -20,7 +20,7 @@ const FEATURES = [
     title: "Managed Postgres, colocated with your Lakehouse.",
     description:
       "Provision with the CLI, connect like any Postgres. Instant branching, scales to zero, and change data feed to Unity Catalog.",
-    href: "/product/data-lakehouse",
+    href: "/product/lakebase",
     visual: "lakebase",
   },
   {
